% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hawkes_exp_bd_slow}
\alias{hawkes_exp_bd_slow}
\title{Simulate Homogeneous non-Markov (Hawkes) Birth-Death Process}
\usage{
hawkes_exp_bd_slow(mu, alpha, beta, tmax, seed, maxN = 10000L)
}
\arguments{
\item{alpha}{mean number of offspring for each particle}

\item{beta}{rate parameter of exponentially distributed particle lifespan}

\item{tmax}{maximum simulation time}

\item{seed}{a seed (if set to 0, will draw a "random" seed from the system time)}

\item{maxN}{if population exceeds this number kill the simulation (useful for preventing runaway processes)}

\item{baseline}{immigration rate (constant; corresponds to homogeneous Poisson immigrant arrivals)}
}
\description{
The slow version here simulates as a marked Hawkes process, by sampling the immigration process as well as 
the marked descendent process.
}
\examples{
\dontrun{

plot(stepfun(x = out$t,y = c(0,out$X)),xlim=c(0,ceiling(max(out$t))))
}
}
