% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_bd}
\alias{poisson_bd}
\title{Simulate Homogeneous Markov (Poisson) Birth-Death Process}
\usage{
poisson_bd(lambda, mu, tmax, seed, maxN = 10000L)
}
\arguments{
\item{lambda}{birth intensity is proportional to lambda}

\item{mu}{death intensity is proportional to mu}

\item{tmax}{maximum simulation time}

\item{seed}{a seed (if set to 0, will draw a "random" seed from the system time)}

\item{maxN}{if population exceeds this number kill the simulation (useful for preventing runaway processes)}
}
\description{
Simulate Homogeneous Markov (Poisson) Birth-Death Process
}
\examples{
\dontrun{
out <- poisson_bd(lambda = 1.1,mu = 1.1,tmax = 30,seed = 42L)
plot(stepfun(x = out$t,y = c(0,out$X)),xlim=c(0,ceiling(max(out$t))))
}
}
