% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_bd_poisson}
\alias{simulate_bd_poisson}
\title{Simulate Birth-Death Cladogenesis Process on Tree}
\usage{
simulate_bd_poisson(seed, lambda, mu, duration, maxN, out)
}
\description{
Simulate Birth-Death Cladogenesis Process on Tree
}
\examples{
\dontrun{
simulate_bd_poisson(seed = 1948L,lambda = 1.5,mu = 0.95,duration = 5,maxN = 1e3L,out = "/Users/slwu89/Desktop/out.nwk")
MyTree <- ape::read.tree("/Users/slwu89/Desktop/out.nwk")
ggtree::ggtree(MyTree)
}
}
